Rework review modal to form and add tags
